refactor(whiteboards): Move save toolbar to a separate component
